// apps/api/src/modules/invitation/enums/invitation.enum.ts
export enum ENUM_INVITATION_STATUS {
    PENDING = 'PENDING',
    ACCEPTED = 'ACCEPTED',
    EXPIRED = 'EXPIRED',
    REVOKED = 'REVOKED',
}

export enum ENUM_INVITATION_STATUS_CODE_ERROR {
    NOT_FOUND = 5800,
    ALREADY_ACCEPTED = 5801,
    EXPIRED = 5802,
    REVOKED = 5803,
    INVALID_TOKEN = 5804,
    ALREADY_MEMBER = 5805,
    CANNOT_INVITE_OWNER = 5806,
}